+++ kupid.ts
@@ -1,5 +1,5 @@
-import type { NoticeInfo, KupidType } from "../type.d.ts";
-import { getToken, trim } from "../utils.ts";
+import type { NoticeInfo, KupidType } from "./types.ts";
+import { getToken, trim } from "./utils.ts";
 
 interface GrwResponse {
   grwSessionId: string;
@@ -115,11 +115,6 @@ function makeFilePathPublic(html: string) {
   });
 }
 
-/**
- * @param id KUPID id
- * @param password KUPID password
- * @returns KUPID '일반공지' 공지사항 10개의 HTML string Array를 반환합니다.
- */
 async function getNoticesFromKupid(id: string, password: string): Promise<string[]> {
   const { token, sessionId } = await getToken(id, password);
   const { grwSessionId, html } = await getNoticeListPage(token, sessionId, "11");
@@ -134,11 +129,6 @@ async function getNoticesFromKupid(id: string, password: string): Promise<string
   );
 }
 
-/**
- * @param id KUPID id
- * @param password KUPID password
- * @returns KUPID '학사일정' 공지사항 10개의 HTML string Array를 반환합니다.
- */
 async function getSchedulesFromKupid(id: string, password: string): Promise<string[]> {
   const { token, sessionId } = await getToken(id, password);
   const { grwSessionId, html } = await getNoticeListPage(token, sessionId, "89");
@@ -152,11 +142,6 @@ async function getSchedulesFromKupid(id: string, password: string): Promise<stri
   );
 }
 
-/**
- * @param id KUPID id
- * @param password KUPID password
- * @returns KUPID '학사일정' 공지사항 10개의 HTML string Array를 반환합니다.
- */
 async function getScholarFromKupid(id: string, password: string): Promise<string[]> {
   const { token, sessionId } = await getToken(id, password);
   const { grwSessionId, html } = await getNoticeListPage(token, sessionId, "88");
@@ -170,10 +155,6 @@ async function getScholarFromKupid(id: string, password: string): Promise<string
   );
 }
 
-/**
- * @param html getNoticesFromKupid 함수에서 반환된 HTML string
- * @returns 공지사항의 제목, 작성자, 게시일자, KUPID 내부 id, public URL, HTML table body의 내용을 반환합니다.
- */
 function parseNoticeInfo(html: string): NoticeInfo {
   const tableRows = html.split("<tr>").slice(1);
   tableRows[tableRows.length - 1] = tableRows[tableRows.length - 1].split("</tr>")[0];
@@ -204,10 +185,7 @@ function parseNoticeInfo(html: string): NoticeInfo {
     url,
   };
 }
-/**
- * @param html getSchedulesFromKupid 함수에서 반환된 HTML string
- * @returns 공지사항의 제목, 작성자, 게시일자, KUPID 내부 id, public URL, HTML table body의 내용을 반환합니다.
- */
+
 function parseScheduleInfo(html: string): NoticeInfo {
   const tableRows = html.split("<tr>").slice(1);
   tableRows[tableRows.length - 1] = tableRows[tableRows.length - 1].split("</tr>")[0];
@@ -239,10 +217,6 @@ function parseScheduleInfo(html: string): NoticeInfo {
   };
 }
 
-/**
- * @param html getScholarFromKupid 함수에서 반환된 HTML string
- * @returns 공지사항의 제목, 작성자, 게시일자, KUPID 내부 id, public URL, HTML table body의 내용을 반환합니다.
- */
 function parseScholarInfo(html: string): NoticeInfo {
   const tableRows = html.split("<tr>").slice(1);
   tableRows[tableRows.length - 1] = tableRows[tableRows.length - 1].split("</tr>")[0];
@@ -274,14 +248,8 @@ function parseScholarInfo(html: string): NoticeInfo {
   };
 }
 
-/**
- *
- * @param id KUPID id
- * @param password KUPID password
- * @param type 장학공지 Scholar, 일반공지 Notice, 학사일정 Schedule
- * @returns KUPID의 해당 카테고리 공지사항 최근 10개의 raw HTML string array를 반환합니다.
- */
-export async function fetchKupidNotices(
+/** notice content from KUPID. */
+export function fetchKupidNotices(
   id: string,
   password: string,
   type: KupidType
@@ -298,13 +266,7 @@ export async function fetchKupidNotices(
   }
 }
 
-/**
- *
- * @param id KUPID id
- * @param password KUPID password
- * @param type 장학공지 Scholar, 일반공지 Notice, 학사일정 Schedule
- * @returns KUPID의 해당 카테고리 공지사항 최근 10개의 content HTML, 제목, 작성자, 게시일자, KUPID 내부 id, public URL을 반환합니다.
- */
+/** return metadata and notice content from KUPID. */
 export async function fetchParsedKupidNotices(
   id: string,
   password: string,
