--- src/type-b/index.ts
@@ -1,2 +1,0 @@
-export * from "./info.ts";
-export * from "./media.ts";
