+++ test.html
@@ -1,0 +1,12 @@
+src="/images/groupware/btn_file.gif" /images/groupware/btn_file.gif
+src="https://biz.korea.ac.kr/images/groupware/btn_file.gif" https://biz.korea.ac.kr/images/groupware/btn_file.gif
+href="/files/groupware/upload/202102/202101121703000521_%ED%95%B5%EC%8B%AC%EA%B5%90%EC%96%91_%EC%9D%B8%EC%A0%95_%EC%8B%A0%EC%B2%AD%EC%84%9C(%EB%B3%80%EA%B2%BD%EC%96%91%EC%8B%9D).hwp" /files/groupware/upload/202102/202101121703000521_%ED%95%B5%EC%8B%AC%EA%B5%90%EC%96%91_%EC%9D%B8%EC%A0%95_%EC%8B%A0%EC%B2%AD%EC%84%9C(%EB%B3%80%EA%B2%BD%EC%96%91%EC%8B%9D).hwp
+
+
+                <img src="https://biz.korea.ac.kr/images/groupware/btn_file.gif" alt="첨부파일" style="cursor:pointer;" onclick="toggleLayerDisplay('fileLayer1');">
+                <img src="https://biz.korea.ac.kr/images/groupware/btn_file.gif" alt="첨부파일" style="cursor:pointer;" onclick="toggleLayerDisplay('fileLayer1');">
+                                             <th style="min-width: 60px;">첨부파일</th>
+                                             <td colspan="5">
+
+                                             <p><a href="https://biz.korea.ac.kr/files/groupware/upload/202102/202101121703000521_%ED%95%B5%EC%8B%AC%EA%B5%90%EC%96%91_%EC%9D%B8%EC%A0%95_%EC%8B%A0%EC%B2%AD%EC%84%9C(%EB%B3%80%EA%B2%BD%EC%96%91%EC%8B%9D).hwp" title="핵심교양_인정_신청서(변경양식).hwp">핵심교양_인정_신청서(변..</a> [크기 : 35 KB / 다운수 : 430]</p>
+
